requirements.txt
flask
flask-cors
python-dotenv
langchain
qdrant-client
pypdf2
numpy
tiktoken
openai
requests
uvicorn
fastapi
langchain-community
